requirements.txt
anthropic>=0.42.0
anthropic_bedrock>=0.8.0
docx>=0.2.4
matplotlib>=3.5.0
mistralai>=1.2.5
numpy>=1.21.0
openai>=1.58.1
pandas>=1.3.0
pdfplumber>=0.10.3
python-dotenv>=1.0.1
Requests>=2.32.3
scipy>=1.7.0
seaborn>=0.13.0
statsmodels>=0.14.0
tiktoken>=0.8.0
PyPDF2>=3.0.1
pytesseract>=0.3.13
